DataPipeline/requirements.txt
sentence-transformers
google-cloud-bigquery
google-cloud-storage
pandas
numpy
chromadb
tf-keras
langchain-google-vertexai
streamlit